Configuring routes
- /api/addUser goes to routes/users/users-route
- /api/trigger-requests goes to /routes/test/trigger-requests-route

- Created some folder structure with modules so we move away from coding in one file
- Adding utils for helper functions like string validation
- Adding utils/middleware for all the helper middleware, like param validation
- Moving graphql call to api/leetcode
