+++ backend/src/api/leetcode.ts
@@ -1,0 +1,36 @@
+import axios from 'axios';
+import { isValidUsername } from '../utils/utils';
+
+// Function to make GraphQL requests for a specific user ID
+export async function makeGraphQLRequest(userId: string) {
+    if(!isValidUsername(userId)) return;
+    try {
+      // Make the GraphQL request
+      const response = await axios.post('https://leetcode.com/graphql', {
+        query: `
+        { 
+          matchedUser(username: "${userId}") 
+          {
+              username
+              submitStats: submitStatsGlobal 
+              {
+                  acSubmissionNum 
+                  {
+                      difficulty
+                      count
+                      submissions
+                  }
+              }
+          }
+      }
+        `
+      });
+  
+      // Handle the response
+      console.log(`User ID: ${userId}`);
+      console.log(JSON.stringify(response.data.data.matchedUser));
+      return response.data.data.matchedUser;
+    } catch (error) {
+      console.error(`Error fetching data for user ID ${userId}:`, error);
+    }
+}
