+++ backend/src/model/users.ts
@@ -1,0 +1,4 @@
+const userIds = new Set<string>();
+
+export const getUserIDs = () => userIds;
+export const addUserID = (userID: string) => userIds.add(userID);
